pyproject.toml
[tool.poetry]
name = "bloblog"
version = "0.1.0"
description = "Synchronize local files with S3, store metadata in DynamoDB."
authors = ["Hong E <hong.e@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/bloblog"
repository = "https://github.com/yourusername/bloblog"
keywords = ["aws", "s3", "dynamodb", "sync", "automation"]

# Instruct Poetry to look for the 'bloblog' package inside 'src'
packages = [
  { include = "bloblog", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = ">=1.35.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
moto = "^5.0.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.6.0"

[tool.poetry.scripts]
bloblog = "bloblog.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Optional development tools configuration
[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
